Refactored to implement commands and namespaces.

+++ cycle-4/lib/busybox.py
@@ -9,16 +9,20 @@ from .util import download_file, make, make_config, exec
 
 class BusyBoxManager:
 
-    def __init__(self, version, builds_dir) -> None:
+    def __init__(self, version: str, builds_dir: str, namespace_dir: str) -> None:
         self._version = version
-        self._build_dir = os.path.join(builds_dir, "busybox")
+        self._namespace_dir = namespace_dir
+
         filename = f"busybox-{version}.tar.bz2"
+
+        self._download_dir = os.path.join(builds_dir, "busybox")
         self._download_url = f"https://busybox.net/downloads/{filename}"
-        self._download_file = os.path.join(self._build_dir, filename)
-        self._source_dir = os.path.join(self._build_dir, f"busybox-{version}")
+        self._download_file = os.path.join(self._download_dir, filename)
+       
+        self._source_dir = os.path.join(self._namespace_dir, f"busybox-{version}")
         
     def download_busybox(self) -> None:
-        os.makedirs(self._build_dir, exist_ok=True)
+        os.makedirs(self._download_dir, exist_ok=True)
 
         if not os.path.exists(self._download_file):
             print(f"Downloading Busy Box {self._version} from {self._download_url}")
@@ -31,7 +35,7 @@ class BusyBoxManager:
         if not os.path.exists(self._source_dir):
             print(f"Extracting Busy Box {self._version} to: {self._source_dir}")
             with tarfile.open(download_file) as f:
-                f.extractall("build/busybox")
+                f.extractall(self._namespace_dir)
         else:
             print(f"Busy Boxy {self._version} already extracted")
 
@@ -46,9 +50,9 @@ class BusyBoxManager:
             for line in lines:
                 sources.write(re.sub(r'^# CONFIG_STATIC is not set$', 'CONFIG_STATIC=y', line))
 
-    def make_busybox(self):
+    def make_busybox(self, rebuild):
         busy_box_build = os.path.join(self._source_dir, "_install")
-        if not os.path.exists(busy_box_build):
+        if not os.path.exists(busy_box_build) or rebuild:
             print("building busy box")
             make(self._source_dir)
             exec("make install", self._source_dir)
@@ -72,20 +76,21 @@ class BusyBoxManager:
         with open(os.path.join(root_fs_build_dir, "init"), "w") as init_file:
             init_file.write(init)
 
-    def build_root_fs(self, root_fs_dir) -> str:
+    def build_root_fs(self, root_fs_dir, rebuild) -> str:
 
-        root_fs_build_dir = os.path.join("build", root_fs_dir)
+        root_fs_build_dir = os.path.join(self._namespace_dir, root_fs_dir)
 
-        shutil.rmtree(root_fs_build_dir, ignore_errors=True)
+        if not os.path.exists(root_fs_build_dir) or rebuild:
+            shutil.rmtree(root_fs_build_dir, ignore_errors=True)
 
-        os.makedirs(root_fs_build_dir, exist_ok=True)
-        
-        exec(f"cp -a {self._source_dir}/_install/* {root_fs_build_dir}")
-        exec("rm linuxrc", root_fs_build_dir)
-        exec("mkdir proc sys dev tmp", root_fs_build_dir)
-        self.write_init(1000, root_fs_build_dir)
-        exec(f"chmod +x {root_fs_build_dir}/init")
+            os.makedirs(root_fs_build_dir, exist_ok=True)
+            
+            exec(f"cp -a {self._source_dir}/_install/* {root_fs_build_dir}")
+            exec("rm linuxrc", root_fs_build_dir)
+            exec("mkdir proc sys dev tmp", root_fs_build_dir)
+            self.write_init(1000, root_fs_build_dir)
+            exec(f"chmod +x {root_fs_build_dir}/init")
 
-        exec(f"find . | cpio -ov --format=newc > ../initramfs.cpio 2>/dev/null", root_fs_build_dir)
+            exec(f"find . | cpio -ov --format=newc > ../initramfs.cpio 2>/dev/null", root_fs_build_dir)
 
-        return "build/initramfs.cpio"
+        return os.path.join(self._namespace_dir, "initramfs.cpio")
